src/main/java/com/mestizo/controller/UserController.java
package com.mestizo.controller;

import com.mestizo.model.User;
import com.mestizo.model.dto.UserDto;
import com.mestizo.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("user")
public class UserController {

    @Autowired
    private UserService userService;

    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    public List<UserDto> findAll(){
        return userService.findAll();
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public String save(@RequestBody UserDto userDto){
        return userService.save(userDto);
    }

}
